Modify app.rb, book.rb, books.json and main.rb files
