+++ data/books.json
@@ -1,1 +1,1 @@
-[{"publisher":"dha","cover_state":"thy","publish_date":"12/4/2022"}]
+[{"title":null,"publisher":"dha","cover_state":"thy","publish_date":"12/4/2022"},{"title":"jesus","publisher":"dha","cover_state":"kam","publish_date":2022},{"title":"read","publisher":"quin","cover_state":"kam","publish_date":2022}]
